+++ TDEngine2/include/graphics/UI/CInputReceiverComponent.h
@@ -77,6 +77,8 @@ namespace TDEngine2
 			bool mPrevState     : 1;
 			bool mCurrState     : 1;
 			bool mIsIgnoreInput : 1;
+
+			TVector2 mNormalizedInputPosition; ///< Position of an input event that's normalized relative to receiver's rectangle
 	};